shoeuniverse/requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
Django==1.11.9
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.2
django-libsass==0.7
django-modelcluster==3.1
django-taggit==0.22.2
django-treebeard==4.2.0
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.6
libsass==0.13.4
Pillow==5.0.0
pytz==2017.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
wagtail==1.13.1
wagtailgmaps==0.3.1
webencodings==0.5.1
Willow==1.0
